Chnaged all isize and usize to i64

+++ src/client/log.rs
@@ -6,7 +6,7 @@ use crate::{
 impl super::Client {
     pub async fn get_log(
         &self,
-        last_known_id: Option<usize>,
+        last_known_id: Option<i64>,
         log_types: Option<Vec<LogType>>,
     ) -> Result<Vec<LogItem>, Error> {
         let mut url = self.build_url("api/v2/log/main").await?;
@@ -54,7 +54,7 @@ impl super::Client {
         Ok(log)
     }
 
-    pub async fn get_log_peer(&self, last_known_id: Option<usize>) -> Result<Vec<LogPeers>, Error> {
+    pub async fn get_log_peer(&self, last_known_id: Option<i64>) -> Result<Vec<LogPeers>, Error> {
         let mut url = self.build_url("api/v2/log/peers").await?;
         if let Some(id) = last_known_id {
             url.set_query(Some(&format!("last_known_id={}", id)));
